+++ packages/react/src/models/RouterOptions.ts
@@ -1,0 +1,4 @@
+export interface RouterOptions {
+  as?: string;
+  unmount?: boolean;
+}